Change GlobalQuote types to index.ts file
