--- src/@types/GlobalQuote.ts
@@ -1,12 +1,0 @@
-export type GlobalQuoteResults = {
-    symbol: string;
-    open: string;
-    high: string;
-    low: string;
-    price: string;
-    volume: string;
-    latestTradingDay: string;
-    previousClose: string;
-    change: string;
-    changePercent: string;
-};
